requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
PyQt6>=6.6.0
pandas>=2.1.0
pyinstaller>=6.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0